# Add new cache statistics 