@drichmond 2021-12-01
Added a new table for cache and DRAM bandwidth statistics. 

NB: nan are inserted because it's not possible to compute the breakdown.

```
----------------------------------------------------------------------------------------------------------------------------
DRAM and Cache ($) Bandwidth Statistics                                                                                     
----------------------------------------------------------------------------------------------------------------------------
  Type    DRAM Rd BW (B/Cyc)    DRAM Wr BW (B/Cyc)    $ Rd BW (B/Cyc)    $ Rd Util. (%)    $ Wr BW (B/Cyc)    $ Wr Util. (%)
------  --------------------  --------------------  -----------------  ----------------  -----------------  ----------------
 Store                  0.03                   nan               0.00              0.00               0.03              0.02
  Load                  0.11                  0.00               0.11              0.09               0.00              0.00
Atomic                  0.00                   nan               0.00              0.00               0.00              0.00
 Total                  0.14                  0.02               0.11              0.09               0.03              0.02
----------------------------------------------------------------------------------------------------------------------------
```